+++ app/src/main/res/layout/fragment_movie_list.xml
@@ -8,7 +8,7 @@
         android:id="@+id/main"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".view.MovieListFragment">
+        tools:context=".view.movies.MovieListFragment">
 
 
         <androidx.recyclerview.widget.RecyclerView
